Add files via upload
CS50 Problem Set 0 - Project name: Blackbird Dash! 

Blackbird Dash is a side scrolling game. The objective is to collect 50 Letters that spell "Hello World", in order to reveal the CS50 duck. At the start of the game, when clicking the green flag, the user is told to press the space bar which initiates the "Dash".